% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_list_all}
\alias{BIEN_list_all}
\title{Extract a list of all species in the BIEN database.}
\usage{
BIEN_list_all(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
Dataframe containing a list of all species in the BIEN database.
}
\description{
BIEN_list_all produces a list of all species in the BIEN database.
}
\examples{
\dontrun{
species_list<-BIEN_list_all()}
}
\seealso{
Other list functions: 
\code{\link{BIEN_list_country}()},
\code{\link{BIEN_list_county}()},
\code{\link{BIEN_list_spatialpolygons}()},
\code{\link{BIEN_list_state}()}
}
\concept{list functions}
