% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_metadata_data_dictionaries}
\alias{BIEN_metadata_data_dictionaries}
\title{Download data dictionaries}
\usage{
BIEN_metadata_data_dictionaries(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
A list containing data.frames containing different data dictionaries.
"Tables" contains information on the contents of tables.
"Columns" contains information on the contents of columns within tables.
"Values" contains information on some of the values that are used to populated columns, particularly where these may be unclear.
"rbien" contains information on some of the values returned by the BIEN R package.
}
\description{
BIEN_metadata_data_dictionaries downloads the data dictionaries for the BIEN database.
}
\examples{
\dontrun{
BIEN_metadata_data_dictionaries()}
}
\seealso{
Other metadata functions: 
\code{\link{BIEN_metadata_citation}()},
\code{\link{BIEN_metadata_database_version}()},
\code{\link{BIEN_metadata_list_political_names}()},
\code{\link{BIEN_metadata_match_data}()},
\code{\link{BIEN_plot_metadata}()},
\code{\link{BIEN_ranges_list}()}
}
\concept{metadata functions}
\keyword{internal}
