% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_metadata_database_version}
\alias{BIEN_metadata_database_version}
\title{Download the current BIEN database version and release date}
\usage{
BIEN_metadata_database_version(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
A data frame containing the current version number and release date for the BIEN database.
}
\description{
BIEN_metadata_database_version downloads the current version number and release date for the BIEN database.
}
\examples{
\dontrun{
BIEN_metadata_database_version()}
}
\seealso{
Other metadata functions: 
\code{\link{BIEN_metadata_citation}()},
\code{\link{BIEN_metadata_list_political_names}()},
\code{\link{BIEN_metadata_match_data}()},
\code{\link{BIEN_plot_metadata}()},
\code{\link{BIEN_ranges_list}()}
}
\concept{metadata functions}
