% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_plot_list_datasource}
\alias{BIEN_plot_list_datasource}
\title{List available datasources.}
\usage{
BIEN_plot_list_datasource(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
A vector of available datasources.
}
\description{
BIEN_plot_list_datasource list all plot datasources in the BIEN database.
}
\examples{
\dontrun{
BIEN_plot_list_datasource()}
}
\seealso{
Other plot functions: 
\code{\link{BIEN_plot_country}()},
\code{\link{BIEN_plot_dataset}()},
\code{\link{BIEN_plot_datasource}()},
\code{\link{BIEN_plot_list_sampling_protocols}()},
\code{\link{BIEN_plot_metadata}()},
\code{\link{BIEN_plot_name}()},
\code{\link{BIEN_plot_sampling_protocol}()},
\code{\link{BIEN_plot_sf}()},
\code{\link{BIEN_plot_state}()}
}
\concept{plot functions}
