% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_plot_metadata}
\alias{BIEN_plot_metadata}
\title{Download plot metadata}
\usage{
BIEN_plot_metadata(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
A dataframe containing plot metadata.
}
\description{
BIEN_plot_metadata downloads the plot metadata table.
}
\examples{
\dontrun{
BIEN_plot_metadata()}
}
\seealso{
Other plot functions: 
\code{\link{BIEN_plot_country}()},
\code{\link{BIEN_plot_dataset}()},
\code{\link{BIEN_plot_datasource}()},
\code{\link{BIEN_plot_list_datasource}()},
\code{\link{BIEN_plot_list_sampling_protocols}()},
\code{\link{BIEN_plot_name}()},
\code{\link{BIEN_plot_sampling_protocol}()},
\code{\link{BIEN_plot_sf}()},
\code{\link{BIEN_plot_state}()}

Other metadata functions: 
\code{\link{BIEN_metadata_citation}()},
\code{\link{BIEN_metadata_data_dictionaries}()},
\code{\link{BIEN_metadata_database_version}()},
\code{\link{BIEN_metadata_list_political_names}()},
\code{\link{BIEN_metadata_match_data}()},
\code{\link{BIEN_ranges_list}()}
}
\concept{metadata functions}
\concept{plot functions}
