% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_plot_name}
\alias{BIEN_plot_name}
\title{Download plot data by plot name.}
\usage{
BIEN_plot_name(
  plot.name,
  cultivated = FALSE,
  new.world = NULL,
  all.taxonomy = FALSE,
  native.status = FALSE,
  natives.only = TRUE,
  political.boundaries = FALSE,
  collection.info = FALSE,
  all.metadata = FALSE,
  ...
)
}
\arguments{
\item{plot.name}{A plot name or vector of names. See BIEN_plot_metadata for more information on plots.}

\item{cultivated}{Return cultivated records as well?  Default is FALSE.}

\item{new.world}{NULL (The default) returns global records, TRUE returns only New World, and FALSE only Old World.}

\item{all.taxonomy}{Return all taxonomic information?  This includes the raw data as well as the "scrubbed" data.}

\item{native.status}{Return information on introduction status?  The default value is FALSE. A value of TRUE also returns additional information on introduction status.}

\item{natives.only}{Exclude detected introduced species?  Default is TRUE.}

\item{political.boundaries}{Return information on political boundaries for an observation? The default value is FALSE.}

\item{collection.info}{Return additional information about collection and identification? The default value is FALSE.}

\item{all.metadata}{Should additional plot metadata be returned?  Default is FALSE.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
A dataframe containing all data from the specified plot(s).
}
\description{
BIEN_plot_name downloads all plot data for a set of plot names.
}
\note{
US FIA coordinates have been fuzzed and swapped, for more details see: https://www.fia.fs.fed.us/tools-data/spatial/Policy/index.php

Plot names can be looked up with \code{\link{BIEN_plot_metadata}}.
}
\examples{
\dontrun{
BIEN_plot_name("SR-1")}
}
\seealso{
Other plot functions: 
\code{\link{BIEN_plot_country}()},
\code{\link{BIEN_plot_dataset}()},
\code{\link{BIEN_plot_datasource}()},
\code{\link{BIEN_plot_list_datasource}()},
\code{\link{BIEN_plot_list_sampling_protocols}()},
\code{\link{BIEN_plot_metadata}()},
\code{\link{BIEN_plot_sampling_protocol}()},
\code{\link{BIEN_plot_spatialpolygons}()},
\code{\link{BIEN_plot_state}()}
}
\concept{plot functions}
