% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_ranges_list}
\alias{BIEN_ranges_list}
\title{List available range maps}
\usage{
BIEN_ranges_list(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
A data.frame containing the available species and their associated GIDs.
}
\description{
BIEN_ranges_list a data.frame containing listing all range maps currently available.
}
\examples{
\dontrun{
available_maps<-BIEN_ranges_list()}
}
\seealso{
Other range functions: 
\code{\link{BIEN_ranges_box}()},
\code{\link{BIEN_ranges_genus}()},
\code{\link{BIEN_ranges_intersect_species}()},
\code{\link{BIEN_ranges_load_species}()},
\code{\link{BIEN_ranges_sf}()},
\code{\link{BIEN_ranges_shapefile_to_skinny}()},
\code{\link{BIEN_ranges_skinny_ranges_to_richness_raster}()},
\code{\link{BIEN_ranges_species}()},
\code{\link{BIEN_ranges_species_bulk}()}

Other metadata functions: 
\code{\link{BIEN_metadata_citation}()},
\code{\link{BIEN_metadata_data_dictionaries}()},
\code{\link{BIEN_metadata_database_version}()},
\code{\link{BIEN_metadata_list_political_names}()},
\code{\link{BIEN_metadata_match_data}()},
\code{\link{BIEN_plot_metadata}()}
}
\concept{metadata functions}
\concept{range functions}
