% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_trait_list}
\alias{BIEN_trait_list}
\title{List all available types of trait data}
\usage{
BIEN_trait_list(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
A dataframe containing all currently available types of trait data and details on measurement.
}
\description{
BIEN_trait_list produces a dataframe of all available types of trait data.
}
\examples{
\dontrun{
BIEN_trait_list()}
}
\seealso{
Other trait functions: 
\code{\link{BIEN_trait_country}()},
\code{\link{BIEN_trait_family}()},
\code{\link{BIEN_trait_genus}()},
\code{\link{BIEN_trait_mean}()},
\code{\link{BIEN_trait_species}()},
\code{\link{BIEN_trait_traitbyfamily}()},
\code{\link{BIEN_trait_traitbygenus}()},
\code{\link{BIEN_trait_traitbyspecies}()},
\code{\link{BIEN_trait_traits_per_species}()},
\code{\link{BIEN_trait_trait}()}
}
\concept{trait functions}
