% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_trait_traits_per_species}
\alias{BIEN_trait_traits_per_species}
\title{Count the number of trait observations for each species in the BIEN database}
\usage{
BIEN_trait_traits_per_species(species = NULL, ...)
}
\arguments{
\item{species}{Optional species or vector of species.  If left blank, returns counts for all species.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
Returns a dataframe containing the number of trait records for each species in the BIEN database.
}
\description{
BIEN_trait_traits_per_species downloads a count of the number of records for each trait for each species in the BIEN database.
}
\examples{
\dontrun{
trait_observation_counts<-BIEN_trait_traits_per_species()}
}
\seealso{
Other trait functions: 
\code{\link{BIEN_trait_country}()},
\code{\link{BIEN_trait_family}()},
\code{\link{BIEN_trait_list}()},
\code{\link{BIEN_trait_mean}()},
\code{\link{BIEN_trait_species}()},
\code{\link{BIEN_trait_traitbyfamily}()},
\code{\link{BIEN_trait_traitbygenus}()},
\code{\link{BIEN_trait_traitbyspecies}()},
\code{\link{BIEN_trait_trait}()}
}
\concept{trait functions}
