% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.is_log_or_null}
\alias{.is_log_or_null}
\title{Check that value is logical or null}
\usage{
.is_log_or_null(x)
}
\description{
Helper function to check data format.
}
\examples{
\dontrun{
is_log_or_null(new.world)}
}
\keyword{internal}
