% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.political_check_traits}
\alias{.political_check_traits}
\title{Set query details}
\usage{
.political_check_traits(political.boundaries)
}
\description{
Helper function to set query components.
}
\keyword{internal}
