% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.species_check}
\alias{.species_check}
\title{Set query details}
\usage{
.species_check(species)
}
\arguments{
\item{species}{Single species or vector of species.}
}
\description{
Helper function to set query components.
}
\keyword{internal}
