% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.vfoi_check_stem}
\alias{.vfoi_check_stem}
\title{Set query details}
\usage{
.vfoi_check_stem(native.status, cultivated, natives.only, collection.info)
}
\description{
Helper function to set query components.
}
\keyword{internal}
